requirements.txt
django==2.0
django-bower==5.2.0
django-phonenumber-field==2.0.0
mysqlclient
pillow
